Rename TestQueue2 -> TestQueue
