+++ client/routes/index.js
@@ -7,7 +7,7 @@ import NotFound from '@components/NotFound';
 import { Reports, Report } from '@components/Reports';
 import CandidateReview from '@components/CandidateReview';
 import SignupInstructions from '@components/SignupInstructions';
-import TestQueue2 from '@components/TestQueue2';
+import TestQueue from '@components/TestQueue';
 import TestRun from '@components/TestRun';
 import UserSettings from '@components/UserSettings';
 import CandidateTestPlanRun from '@components/CandidateReview/CandidateTestPlanRun';
@@ -37,7 +37,7 @@ export default () => (
         </ConfirmAuth>
       }
     />
-    <Route exact path="/test-queue" element={<TestQueue2 />} />
+    <Route exact path="/test-queue" element={<TestQueue />} />
     <Route
       exact
       path="/test-plan-report/:testPlanReportId"
